renamed core to raft
