--- core/__init__.py
@@ -1,1 +1,0 @@
-from raft import RAFT
