+++ raft/__init__.py
@@ -1,0 +1,1 @@
+from raft.raft import RAFT
