+++ raft/raft.py
@@ -3,10 +3,10 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from update import BasicUpdateBlock, SmallUpdateBlock
-from extractor import BasicEncoder, SmallEncoder
-from corr import CorrBlock, AlternateCorrBlock
-from utils.utils import bilinear_sampler, coords_grid, upflow8
+from raft.update import BasicUpdateBlock, SmallUpdateBlock
+from raft.extractor import BasicEncoder, SmallEncoder
+from raft.corr import CorrBlock, AlternateCorrBlock
+from raft.utils.utils import bilinear_sampler, coords_grid, upflow8
 
 try:
     autocast = torch.cuda.amp.autocast
